feat: delegate simulation classes discovery to commons and allow simulation start override
Motivation:
Allow to start a simulation configured with a non-discovered class name, by overriding the value based on plugin configuration or by client input.

Modifications:
* Delegate simulations discovery to commons library
* Pass configured class name along redefined commons method signatures
* Handle introduced exceptions
* Split up EnterpriseSettings
* Take advantage of common interface between batch and interactive
  enterprise plugin

Ref: MISC-348
