+++ build.sbt
@@ -10,7 +10,7 @@ githubPath := "gatling/gatling-sbt-plugin"
 libraryDependencies ++= Seq(
   "org.scalatest"     %% "scalatest"                         % "3.2.11" % Test,
   "org.zeroturnaround" % "zt-zip"                            % "1.15",
-  "io.gatling"         % "gatling-enterprise-plugin-commons" % "1.0.3-M1"
+  "io.gatling"         % "gatling-enterprise-plugin-commons" % "1.1.1"
 )
 
 scriptedLaunchOpts := {
@@ -34,6 +34,11 @@ gatlingDevelopers := Seq(
   GatlingDeveloper(
     "sbrevet@gatling.io",
     "Sébastien Brevet",
+    true
+  ),
+  GatlingDeveloper(
+    "tpetillot@gatling.io",
+    "Thomas Petillot",
     true
   )
 )
