+++ src/main/scala/io/gatling/sbt/settings/gatling/ErrorAlreadyLoggedException.scala
@@ -1,0 +1,23 @@
+/*
+ * Copyright 2011-2022 GatlingCorp (https://gatling.io)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.gatling.sbt.settings.gatling
+
+import sbt.FeedbackProvidedException
+
+// MessageOnlyException seems to be logged multiple times when throw in an inner task
+// To prevent that, we explicitly log the error and throw a FeedbackProvidedException
+private[gatling] object ErrorAlreadyLoggedException extends FeedbackProvidedException
